+++ photutils/psf/tests/test_photometry.py
@@ -101,7 +101,7 @@ def test_invalid_inputs():
 
     match = 'localbkg_estimator must be a LocalBackground instance'
     localbkg = MMMBackground()
-    with pytest.raises(ValueError, match=match):
+    with pytest.raises(TypeError, match=match):
         _ = PSFPhotometry(model, 1, localbkg_estimator=localbkg)
 
     match = 'aperture_radius must be a strictly-positive scalar'
@@ -109,8 +109,8 @@ def test_invalid_inputs():
         with pytest.raises(ValueError, match=match):
             _ = PSFPhotometry(model, 1, aperture_radius=radius)
 
-    match = 'grouper must be a SourceGrouper instance'
-    with pytest.raises(ValueError, match=match):
+    match = "'grouper' must be a callable object"
+    with pytest.raises(TypeError, match=match):
         _ = PSFPhotometry(model, (5, 5), grouper=1)
 
     match = 'data must be a 2D array'
@@ -1333,7 +1333,7 @@ def test_flag32_parameter_at_bounds():
 
 
 def test_psf_photometry_methods(test_data):
-    data, error, sources = test_data
+    data, error, _ = test_data
 
     psf_model = CircularGaussianPRF(flux=1, fwhm=2.7)
     fit_shape = (5, 5)
@@ -1823,7 +1823,7 @@ def test_qfit_cfit_with_different_errors(test_data):
     """
     Test qfit and cfit with different error values.
     """
-    data, error, sources = test_data
+    data, error, _ = test_data
 
     psf_model = CircularGaussianPRF(flux=1, fwhm=2.7)
     fit_shape = (5, 5)